+++ src/build32.sh
@@ -8,5 +8,5 @@ echo ./gen mem.s
 ./gen mem.s
 fi
 
-echo gcc -m32 -g3 cImpls.c lisp.c cell.c mem32.c math.c -o lisp32
-gcc -m32 -g3 cImpls.c lisp.c cell.c mem32.c math.c -o lisp32
+echo gcc -m32 -g3 flow.c lisp.c cell.c mem32.c math.c -o lisp32
+gcc -m32 -g3 flow.c lisp.c cell.c mem32.c math.c -o lisp32
